+++ contrib/localnet/orchestrator/smoketest/smoketests/test_erc20_refund.go
@@ -32,6 +32,7 @@ func TestERC20DepositAndCallRefund(sm *runner.SmokeTestRunner) {
 
 	// There is no liquidity pool, therefore the cctx should abort
 	cctx := utils.WaitCctxMinedByInTxHash(sm.Ctx, inTxHash, sm.CctxClient, sm.Logger, sm.CctxTimeout)
+	sm.Logger.CCTX(*cctx, "deposit")
 	if cctx.CctxStatus.Status != types.CctxStatus_Aborted {
 		panic(fmt.Sprintf("expected cctx status to be Aborted; got %s", cctx.CctxStatus.Status))
 	}